Ignore Fated Affix Chaotic Essence
The Chaotic Essence fated affix spawns an orb that you click on.

When you do, it becomes hostile and spawns 2 small adds (Chaotic Motes). The small adds double when killed, and the big add does nothing. Damaging the Chaotic Essence does not provide any buffs/throughput to the raid. WCL already ignores it on logs.