White Box Error Fix
Users were getting a white box window which was due to a skin not being applied after loading instance_defaults.

This change makes instance_defaults.skin set to "no skin".

On RestauraJanela:
Removed an extra rows_fit_in_window nil, and set the correct value.
Checked if the skin is "no skin" and if so, let the user know the skin wasn't found, and reset to default.
Moved InstanceColor to below this check as it relies on a valid skin.

In NovaInstancia, moved the default ChangeSkin to before AtivarInstancia as that requires a valid skin.