Fix all issues related to new textL_offset and text_yoffset
With the new additions of Text Y Offset and the left text offset. The indexes to disable/hide were not updated properly and thus you could not edit Per Second, percent if multifontstring was enabled. And could not even see the seperator box.

A lot of SetPoint calls were not properly updated with the new anchor settings